requirement.txt
asgiref==3.5.2
asyncio==3.4.3
beartype==0.11.0
bidict==0.22.0
black==22.10.0
cachelib==0.9.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
decorator==4.4.2
distlib==0.3.6
export==0.2.0
filelock==3.8.0
Flask==2.2.2
Flask-Session==0.4.0
Flask-SocketIO==5.3.2
flet==0.1.65
idna==3.4
imageio==2.22.2
imageio-ffmpeg==0.4.7
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
moviepy==1.0.3
mypy-extensions==0.4.3
mysql-connector-python==8.0.31
numpy==1.23.4
oauthlib==3.2.2
pafy==0.5.5
pandas==1.5.0
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
proglog==0.1.10
protobuf==3.20.1
PyAudio==0.2.12
pycodestyle==2.9.1
pydub==0.25.1
PyMySQL==1.0.2
pynput==1.7.6
python-dateutil==2.8.2
python-engineio==4.3.4
python-socketio==5.7.2
python-vlc==3.0.16120
pytz==2022.4
redcmd==1.2.10
redlib==1.5.7
repath==0.9.0
requests==2.28.1
scipy==1.9.3
six==1.16.0
socketIO-client==0.7.2
subprocess.run==0.0.8
tomli==2.0.1
tqdm==4.64.1
urllib3==1.26.12
virtualenv==20.16.7
vlc-ctrl==1.0.8
watchdog==2.1.9
websocket-client==1.4.2
websockets==10.4
Werkzeug==2.2.2
youtube-dl==2020.12.2
ytmusicapi==0.24.0
zipp==3.10.0
zope.interface==5.5.0
